# Use gauss_elimination_interval to solve A*b = x in contractors