--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,2 @@
 docs/build/
 docs/site/
-benchmark/tune.json