--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 docs/build/
 docs/site/
+benchmark/tune.json