--- a/src/IntervalRootFinding.jl
+++ b/src/IntervalRootFinding.jl
@@ -38,19 +38,20 @@ const D = derivative
 
 const where_bisect = IntervalArithmetic.where_bisect ## 127//256
 
+include("complex.jl")
+include("slopes.jl")
+include("linear_eq.jl")
+include("quadratic.jl")
 
 include("root_object.jl")
 
-include("newton.jl")
-include("krawczyk.jl")
-
-include("complex.jl")
 include("contractors.jl")
 include("roots.jl")
+
+# Old stuff
+include("newton.jl")
+include("krawczyk.jl")
 include("newton1d.jl")
-include("quadratic.jl")
-include("linear_eq.jl")
-include("slopes.jl")
 
 
 gradient(f) = X -> ForwardDiff.gradient(f, X[:])