@dpsanders 2019-03-07
From what I've read, it's recommended to use interval Gauss-Seidel. Probably we should change `\` to use that.